# api-doc-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "datum",
    "description": "datum crm system",
    "termsOfService": "http://swagger.io/terms/",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/CLINIC/{clinicId}/employee/{employeeId}/department/{departmentId}/employee/{id}/schedule": {
      "put": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "clinicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Schedule"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          }
        }
      }
    },
    "/CLINIC/{clinicId}/employee/{employeeId}/department/{departmentId}/employee/{id}/post/{postId}": {
      "put": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "updateEmployee_1",
        "parameters": [
          {
            "name": "clinicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
...
    "/CLINIC/{clinicId}/employee/{employeeId}/department/{departmentId}/employee": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "getEmployee_1",
        "parameters": [
          {
            "name": "clinicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "createEmployee",
        "parameters": [
          {
            "name": "clinicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/CLINIC/{clinicId}/employee/{employeeId}/department/{departmentId}/employee/owner": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "createEmployeeByOwner",
        "parameters": [
          {
            "name": "clinicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/CLINIC/{clinicId}/employee/{employeeId}/department/{departmentId}/employee/{id}": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "getEmployee_2",
        "parameters": [
          {
            "name": "clinicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "employee-controller"
        ],
        "operationId": "deleteEmployee",
        "parameters": [
          {
            "name": "clinicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PersonDTO": {
        "required": [
          "firstname"
        ],
        "type": "object",
        "properties": {
          "phone": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            }
          },
          "firstname": {
            "type": "string",
            "example": "Үсөн"
          },
          "surname": {
            "type": "string",
            "example": "Асанов"
          },
          "patronymic": {
            "type": "string",
            "example": "Жаңыбаевич"
          },
          "address": {
            "type": "string",
            "example": "пр. Чуй 205, Бишкек, Кыргызстан"
          },
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          },
          "country": {
            "type": "string",
            "example": "KGZ"
          },
          "family": {
            "type": "boolean",
            "example": true
          },
          "male": {
            "type": "boolean",
            "example": true
          },
          "issue": {
            "type": "string",
            "format": "date",
            "example": "2019-12-31"
          },
          "expiry": {
            "type": "string",
            "format": "date",
            "example": "2029-12-31"
          },
          "authority": {
            "type": "string",
            "example": "MKK"
          },
          "birthDay": {
            "type": "string",
            "format": "date",
            "example": "1991-11-21"
          },
          "passportSeries": {
            "type": "string",
            "example": "AN"
          },
          "passportNumber": {
            "type": "integer",
            "format": "int32",
            "example": 1234567
          },
          "pin": {
            "type": "integer",
            "example": 20101199100001
          }
        }
      },
      "Phone": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string",
            "example": "+996 777 123456"
          },
          "whatsapp": {
            "type": "boolean"
          },
          "telegram": {
            "type": "boolean"
          },
          "viber": {
            "type": "boolean"
          }
        }
      },
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "phone": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            }
          },
          "firstname": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "patronymic": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "male": {
            "type": "boolean"
          },
          "family": {
            "type": "boolean"
          },
          "issue": {
            "type": "string",
            "format": "date"
          },
          "expiry": {
            "type": "string",
            "format": "date"
          },
          "authority": {
            "type": "string"
          },
          "birthDay": {
            "type": "string",
            "format": "date"
          },
          "passportSeries": {
            "type": "string"
          },
          "passportNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pin": {
            "type": "integer"
          }
        }
      },
      "Address": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Phone"
            }
          },
          "country": {
            "type": "string",
            "example": "Кыргызстан"
          },
          "city": {
            "type": "string",
            "example": "Бишкек"
          },
          "region": {
            "type": "string"
          },
          "street": {
            "type": "string",
            "example": "пр. Чүй"
          },
          "house": {
            "type": "integer",
            "format": "int32",
            "example": 205
          },
          "apartment": {
            "type": "integer",
            "format": "int32"
          },
          "latitude": {
            "type": "string",
            "example": "42.877317"
          },
          "longitude": {
            "type": "string",
            "example": "74.600215"
          },
          "iframe": {
            "type": "string",
            "example": "<iframe src=\"https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2923\n.8231570293283!2d74.59588357567849!3d42.87657660231409!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13\n.1!3m3!1m2!1s0x389eb7ddc401db31%3A0xd868494a462beaa9!2z0JHQtdC70YvQuSDQtNC-0LwsINC_0YDQvtGB0\nL8uINCn0YPQuSwg0JHQuNGI0LrQtdC6!5e0!3m2!1sru!2skg!4v1699201561246!5m2!1sru!2skg\"\nwidth=\"600\"\nheight=\"450\"\nstyle=\"border:0;\"\nallowfullscreen=\"\"\nloading=\"lazy\"\nreferrerpolicy=\"no-referrer-when-downgrade\">\n</iframe>\n"
          },
          "link": {
            "type": "string",
            "example": "https://go.2gis.com/gh71q"
          }
        }
      },
...
      "Clinic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "departments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Department"
            }
          },
          "preiskurants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Preiskurant"
            }
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          }
        }
      },
      "CurrencyCode": {
        "type": "string",
        "example": "KGS",
        "enum": [
          "AED",
          "AFN",
          "ALL",
          "AMD",
          "ANG",
          "AOA",
          "ARS",
          "AUD",
          "AWG",
          "AZN",
          "BAM",
          "BBD",
          "BDT",
          "BGN",
          "BHD",
          "BIF",
          "BMD",
          "BND",
          "BOB",
          "BRL",
          "BSD",
          "BTN",
          "BWP",
          "BYR",
          "BZD",
          "CAD",
          "CDF",
          "CHF",
          "CLP",
          "CNY",
          "COP",
          "CRC",
          "CUC",
          "CUP",
          "CVE",
          "CZK",
          "DJF",
          "DKK",
          "DOP",
          "DZD",
          "EGP",
          "ERN",
          "ETB",
          "EUR",
          "FJD",
          "FKP",
          "GBP",
          "GEL",
          "GGP",
          "GHS",
          "GIP",
          "GMD",
          "GNF",
          "GTQ",
          "GYD",
          "HKD",
          "HNL",
          "HRK",
          "HTG",
          "HUF",
          "IDR",
          "ILS",
          "IMP",
          "INR",
          "IQD",
          "IRR",
          "ISK",
          "JEP",
          "JMD",
          "JOD",
          "JPY",
          "KES",
          "KGS",
          "KHR",
          "KMF",
          "KPW",
          "KRW",
          "KWD",
          "KYD",
          "KZT",
          "LAK",
          "LBP",
          "LKR",
          "LRD",
          "LSL",
          "LYD",
          "MAD",
          "MDL",
          "MGA",
          "MKD",
          "MMK",
          "MNT",
          "MOP",
          "MRO",
          "MUR",
          "MVR",
          "MWK",
          "MXN",
          "MYR",
          "MZN",
          "NAD",
          "NGN",
          "NIO",
          "NOK",
          "NPR",
          "NZD",
          "OMR",
          "PAB",
          "PEN",
          "PGK",
          "PHP",
          "PKR",
          "PLN",
          "PYG",
          "QAR",
          "RON",
          "RSD",
          "RUB",
          "RWF",
          "SAR",
          "SBD",
          "SCR",
          "SDG",
          "SEK",
          "SGD",
          "SHP",
          "SLL",
          "SOS",
          "SPL",
          "SRD",
          "STD",
          "SVC",
          "SYP",
          "SZL",
          "THB",
          "TJS",
          "TMT",
          "TND",
          "TOP",
          "TRY",
          "TTD",
          "TVD",
          "TWD",
          "TZS",
          "UAH",
          "UGX",
          "USD",
          "UYU",
          "UZS",
          "VEF",
          "VND",
          "VUV",
          "WST",
          "XAF",
          "XCD",
          "XDR",
          "XOF",
          "XPF",
          "YER",
          "ZAR",
          "ZMW",
          "ZWD"
        ]
      },
      "DayTime": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-24T06:00:00Z"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-24T15:00:00Z"
          }
        }
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "employees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            }
          }
        }
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "MonetaryAmount": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 199.96
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyCode"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "code": {
            "type": "string"
          },
          "en": {
            "type": "string"
          },
          "ru": {
            "type": "string"
          },
          "kg": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Preiskurant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "service": {
            "type": "integer",
            "format": "int64"
          },
          "price": {
            "$ref": "#/components/schemas/MonetaryAmount"
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN",
          "ROOT"
        ]
      },
      "Schedule": {
        "type": "object",
        "properties": {
          "week": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DayTime"
            }
          },
          "secondWeek": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DayTime"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "person": {
            "$ref": "#/components/schemas/Person"
          },
          "employee": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            }
          },
          "clinics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Clinic"
            }
          }
        }
      },
...
      "EmployeeDTO": {
        "required": [
          "post"
        ],
        "type": "object",
        "properties": {
          "post": {
            "type": "string",
            "description": "postCode",
            "example": "DENTIST"
          },
          "user": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          }
        }
      },
      "UserDTO": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "email",
            "example": "user@gmail.com"
          },
          "password": {
            "type": "string"
          },
          "person": {
            "$ref": "#/components/schemas/PersonDTO"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}